add and build docker
