+++ ...txt
@@ -1,0 +1,49 @@
+
+anyio==3.6.2
+certifi==2022.12.7
+charset-normalizer==3.0.1
+click==8.1.3
+colorama==0.4.6
+contourpy==1.0.7
+cycler==0.11.0
+Flask==2.2.2
+fonttools==4.38.0
+greenlet==2.0.1
+h11==0.14.0
+h2==4.1.0
+hpack==4.0.0
+httptools==0.5.0
+hypercorn==0.14.3
+hyperframe==6.0.1
+idna==3.4
+importlib-metadata==6.0.0
+imutils==0.5.4
+itsdangerous==2.1.2
+Jinja2==3.1.2
+kiwisolver==1.4.4
+MarkupSafe==2.1.1
+matplotlib==3.6.3
+numpy==1.24.1
+opencv-python==4.5.3.56
+packaging==23.0
+Pillow==9.4.0
+priority==2.0.
+PyMySQL==1.0.2
+pyparsing==3.0.9
+python-dateutil==2.8.2
+python-dotenv==0.21.0
+PyYAML==6.0
+requests==2.28.2
+six==1.16.0
+sniffio==1.3.0
+SQLAlchemy==1.4.44
+starlette==0.21.0
+toml==0.10.2
+typing-extensions==4.4.0
+urllib3==1.26.14
+uvicorn==0.19.0
+watchfiles==0.18.1
+websockets==10.4
+Werkzeug==2.2.2
+wsproto==1.2.0
+zipp==3.11.0
