+++ .github/workflows/build-image.yml
@@ -7,6 +7,7 @@ on:
     # to the auth route and auth workflow
     branches:
       - "main"
+      - "add_docker"
       
 env:
   REGISTRY: ghcr.io
